Removed BackEnd folder
The purpose of "Adding initial file based on stable" is to merge the repository of [nd-backend](https://github.com/coding-team-sept/nd-backend) to this new repository. However, merging it the current way does not preserve the logs. Instead, `git subtree` will be used to merge the backend to this new repository.